+++ src/main/java/net/more_rpg_classes/util/CustomMethods.java
@@ -1,9 +1,14 @@
 package net.more_rpg_classes.util;
 
+import net.minecraft.entity.AreaEffectCloudEntity;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectInstance;
+import net.minecraft.entity.projectile.ProjectileEntity;
+import net.minecraft.particle.ParticleEffect;
 
+import java.util.Iterator;
 import java.util.List;
 
 public class CustomMethods {
@@ -120,5 +125,41 @@ public class CustomMethods {
             }
         }
         return true;
+    }
+
+    public static void spawnCloudEntity(
+            ParticleEffect particleType, Entity owner, float radiusCloud, int durationSecondsCloud, float radiusGrowthCloud
+            , StatusEffect statusEffect, int durationSecondsStatusEffect, int amplifierStatusEffect) {
+        if (!owner.getWorld().isClient) {
+            List<LivingEntity> list = owner.getWorld().getNonSpectatingEntities(LivingEntity.class, owner.getBoundingBox().expand(4.0, 2.0, 4.0));
+            AreaEffectCloudEntity areaEffectCloudEntity = new AreaEffectCloudEntity(owner.getWorld(), owner.getX(), owner.getY(), owner.getZ());
+            Entity entity = null;
+            if(owner instanceof LivingEntity){
+                entity = owner;
+            }else if(owner instanceof ProjectileEntity projectile){
+                projectile.getOwner();
+            }
+            if (entity instanceof LivingEntity) {
+                areaEffectCloudEntity.setOwner((LivingEntity)entity);
+            }
+            areaEffectCloudEntity.setParticleType(particleType);
+            areaEffectCloudEntity.setRadius(radiusCloud);
+            areaEffectCloudEntity.setDuration(durationSecondsCloud*20);
+            areaEffectCloudEntity.setRadiusGrowth((radiusGrowthCloud - areaEffectCloudEntity.getRadius()) / (float)areaEffectCloudEntity.getDuration());
+            areaEffectCloudEntity.addEffect(new StatusEffectInstance(statusEffect,
+                    durationSecondsStatusEffect*20, amplifierStatusEffect, false, false, true));
+            if (!list.isEmpty()) {
+                Iterator var5 = list.iterator();
+                while(var5.hasNext()) {
+                    LivingEntity livingEntity2 = (LivingEntity)var5.next();
+                    double x = owner.squaredDistanceTo(livingEntity2);
+                    if (x < 16.0) {
+                        areaEffectCloudEntity.setPosition(livingEntity2.getX(), livingEntity2.getY(), livingEntity2.getZ());
+                        break;
+                    }
+                }
+            }
+            owner.getWorld().spawnEntity(areaEffectCloudEntity);
+        }
     }
 }